requirements.txt
streamlit
streamlit-webrtc
numpy
Pillow
av
